pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "docfold"
version = "0.6.12"
description = "Turn any document into structured data. Unified interface for document structuring engines with built-in evaluation."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Mihailorama", email = "mihailorama@gmail.com" },
]
keywords = [
    "document-processing",
    "pdf-extraction",
    "ocr",
    "document-structuring",
    "table-extraction",
    "layout-analysis",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: General",
]

dependencies = []

[project.optional-dependencies]
docling = [
    "docling>=2.0",
]
mineru = [
    "magic-pdf[full]>=0.9",
]
marker = [
    "requests>=2.31",
]
pymupdf = [
    "PyMuPDF>=1.23",
]
paddleocr = [
    "paddleocr>=2.7",
    "paddlepaddle>=2.5",
    "opencv-python>=4.7",
    "Pillow>=10.0",
    "pdf2image>=1.16",
]
tesseract = [
    "pytesseract>=0.3.10",
    "Pillow>=10.0",
    "pdf2image>=1.16",
]
easyocr = [
    "easyocr>=1.7",
    "Pillow>=10.0",
    "pdf2image>=1.16",
]
unstructured = [
    "unstructured[all-docs]>=0.16",
]
llamaparse = [
    "llama-parse>=0.5",
]
mistral-ocr = [
    "mistralai>=1.0",
]
zerox = [
    "py-zerox>=0.0.5",
]
textract = [
    "boto3>=1.28",
]
google-docai = [
    "google-cloud-documentai>=3.0",
]
azure-docint = [
    "azure-ai-documentintelligence>=1.0",
]
nougat = [
    "nougat-ocr>=0.1.17",
    "torch>=2.0",
]
surya = [
    "surya-ocr>=0.6",
    "torch>=2.0",
    "Pillow>=10.0",
]
lightonocr = [
    "transformers>=5.0",
    "torch>=2.0",
    "pypdfium2>=4.0",
    "Pillow>=10.0",
]
firecrawl = [
    "firecrawl-py>=1.0",
]
glm-ocr = [
    "glmocr",
]
glm-ocr-selfhosted = [
    "glmocr[selfhosted]",
]
evaluation = [
    "jiwer>=3.0",          # WER/CER computation
    "numpy>=1.23",
    "scipy>=1.10",         # Kendall's tau for reading order
    "tabulate>=0.9",       # Report tables
    "psutil>=5.9",         # Memory measurement
]
api = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "python-multipart>=0.0.6",
    "pydantic-settings>=2.0",
    "redis>=5.0",
    "arq>=0.25",
]
all = [
    "docfold[docling,mineru,marker,pymupdf,paddleocr,tesseract,easyocr,unstructured,llamaparse,mistral-ocr,glm-ocr,textract,google-docai,azure-docint,nougat,surya,lightonocr,firecrawl,evaluation]",
    # Note: zerox excluded from [all] — py-zerox requires Python 3.11+
    # Install separately: pip install docfold[zerox]
]
test = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "pytest-timeout>=2.1",
    "ruff>=0.4",
    "mypy>=1.8",
    "httpx>=0.27",
]
dev = [
    "docfold[test]",
    # Individual engines installed as needed for local dev
]

[project.urls]
Homepage = "https://github.com/mihailorama/docfold"
Documentation = "https://github.com/mihailorama/docfold/tree/main/docs"
Repository = "https://github.com/mihailorama/docfold"
Issues = "https://github.com/mihailorama/docfold/issues"
Changelog = "https://github.com/mihailorama/docfold/blob/main/CHANGELOG.md"

[project.scripts]
docfold = "docfold.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/docfold"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests that require engine dependencies installed",
    "slow: marks tests that take >30s",
    "gpu: marks tests that require GPU",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_unused_configs = true
disable_error_code = ["import-untyped"]
